+++ src/main/scala/Rt/CustomField.scala
@@ -1,6 +1,7 @@
 package com.benkolera.Rt
 
 import scalaz._
+import scala.concurrent.Future
 import org.joda.time.{DateTime,DateTimeZone}
 import org.joda.time.format.DateTimeFormatter
 import Parser.Read
@@ -16,10 +17,21 @@ case class CustomFieldValue( val string: String ) {
   def toDouble(): String \/ Double = Read.readDouble( toString )
   def toDateTime(dtf:DateTimeFormatter,tz:DateTimeZone): String \/ DateTime =
     Read.readDateTime( dtf , tz )( toString )
+  def toDateTimeRtM(implicit m:Monad[Future]): RtM[String \/ DateTime] = {
+    getConfig.map( c =>
+      toDateTime(c.dateTimeFormatter,c.dateTimeZone)
+    )
+  }
   def toOptDateTime(
     dtf:DateTimeFormatter,tz:DateTimeZone
   ): String \/ Option[DateTime] =
     Read.readOptDateTime( dtf ,tz )( toString )
+
+  def toOptDateTimeRtM(implicit m:Monad[Future]): RtM[String \/ Option[DateTime]] = {
+    getConfig.map( c =>
+      toOptDateTime(c.dateTimeFormatter,c.dateTimeZone)
+    )
+  }
 }
 
 object CustomFieldValue {