+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changes to com.benkolera scala-rt
 =================================
 
+2015-11-19 - 0.8.0 - Ben Kolera
+- Added in a RtM version of getDate from CustomFieldValue
+
 2015-11-11 - 0.7.0 - Ben Kolera
 - Added in a way to grab users from RT.